third_party/blink/web_tests/transforms/usecounter-3d-scene.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/run-after-layout-and-paint.js"></script>
<body>
<script>
const kTransform3dScene = 3867; // from web_feature.mojom

const kScenes = [
  { markup: "<div id='transform'></div>",
    is_3d: false },
  { markup: "<div style='transform: rotateX(30deg)'><div id='transform'></div></div>",
    is_3d: false },
  { markup: "<div style='transform-style: preserve-3d'><div id='transform'></div></div>",
    is_3d: true },
  { markup: "<div style='transform-style: preserve-3d'><div style='transform: rotateZ(30deg)'><div id='transform'></div></div></div>",
    is_3d: false },
  { markup: "<div style='transform-style: preserve-3d'><div style='transform-style: preserve-3d; transform: rotateZ(30deg)'><div id='transform'></div></div></div>",
    is_3d: true }
];

const kTransforms = [
  { value: "perspective(400px)", is_3d: false },
  { value: "rotateX(30deg) perspective(400px)", is_3d: true },
  { value: "perspective(400px) rotateX(30deg)", is_3d: true },
  { value: "rotateY(30deg)", is_3d: true },
  { value: "matrix3d(2, 2, 0, 2, 2, 2, 0, 2, 2, 2, 2, 2, 2, 2, 0, 2)", is_3d: false },
  { value: "Translatez(20px)", is_3d: true },
  { value: "matrix3d(0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0)", is_3d: true },
  { value: "matrix3d(0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0)", is_3d: true },
  { value: "matrix3d(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0)", is_3d: true },
];

for (const scene of kScenes) {
  for (const transform of kTransforms) {
    let iframe, idoc;
    promise_test((test) => {
      return Promise.resolve(true).then(() => {
        iframe = document.createElement("iframe");
        document.body.appendChild(iframe);
        idoc = iframe.contentDocument;
        idoc.body.innerHTML = "<style>#transform { width: 50px; height: 50px; background: blue; }</style>" + scene.markup;
        idoc.getElementById("transform").style.transform = transform.value;
        return new Promise((resolve, reject) => {
          // wait for it to paint
          runAfterLayoutAndPaint(() => { resolve(true); });
        });
      }).then(() => {
        assert_equals(internals.isUseCounted(idoc, kTransform3dScene),
                      scene.is_3d && transform.is_3d);
        document.body.removeChild(iframe);
      });
    }, `kTransform3dScene use counter with transform ${transform.value} and scene ${scene.markup}`);
  }
}
</script>
